Create README.md
In this project i had to make a landing page using html and css, in nuncba bootcamp i saw html and css so far, the landing page is about a coffe shop with wholesale to the public. Im very happy tu show you my first landing. The first of many. See you.